SEQNG-157 Move Dropdown component #time 10m
